10-rbac/rbac_demo/requirements.txt
requests
kubernetes
python-dotenv